Changed TurnByCommand to use a double for amount

+++ src/main/java/frc/robot/commands/DictatorLocator.java
@@ -34,7 +34,7 @@ public class DictatorLocator extends CommandBase {
     isFinished = false;
     timer.start();
     timer.reset();
-    MarkAntony = new TurnByCommand((int)Brutus.getHorizontalOffSet(), GermanicCalvalry, 1);
+    MarkAntony = new TurnByCommand(Brutus.getHorizontalOffSet(), GermanicCalvalry, 1);
   }
 
   @Override
